This speeds up the target generation of QEMU, which is very slow.

* Before: 41 seconds (total time 174 seconds)
* After: ~~29~~ 21 seconds (total time ~~150~~ 134 seconds)