@bonzini 2024-11-20
> The backend changes look very interesting, though also take quite a bit more effort to understand what the code is doing both before and after. :D

Fair enough! I've split the changes more finely, which should be easier to both review and apply individually. If you prefer to have multiple PRs let me know.