@eli-schwartz 2024-12-30
Cherry-picked: e542901af6e30865715d3c3c18f703910a096ec0...bb23db4ee33f7058490b4f59d1a28178db9a9ea7