@bonzini 2024-12-30
Thanks, I reduced the rest to just three that provide the most bang for the buck (and are easier to review than it seems).